Corridors constraint

It is important to maintain connectivity between reserves. However, some areas are more difficult for species to traverse then other areas. As a consequence, even though reserves may be connected, species may not be able to move between reserves if the areas connecting them are barriers to dispersal.

add_corridors_constraint(x, targets, friction)

Arguments

x

ConservationProblem-class object.

targets

numeric targets specifying the minimum flow permitted throughout the solution. If targets is a single number then it is used for each feature. Otherwise, if targets is a vector then it specifies the target for each feature.

friction

RasterStack-class object denoting friction surfaces for each feature. Cells with zero values offer no cost to dispersal and individuals can move through these with ease. Cells with higher values are associated with a cost for traversal. Cells with NA values are complete barriers to dispersal and individuals cannot move through these at all.

Value

ConservationProblem-class object.

Details

This function adds constraints to ensure that corridors connect reserves and that individuals from all species can utilise the corridors. Friction Raster-class objects area used to show each dificult areas are to traverse.

See also

constraints for all the available constraints.

Examples

## Not run: ------------------------------------
# 
# # create basic problem
# p1 <- problem(sim_pu_raster, sim_features) %>%
#   add_minimum_set_objective() %>%
#   add_relative_targets(0.2) %>%
#   add_binary_decision()
# 
# # create problem with added corridor constraints
# p2 <- p1 %>% add_corridors_constraint(0.2, sim_friction_raster)
# 
# # solve problems
# s <- stack(solve(p1), solve(p2))
# 
# # plot solutions
# plot(s, main=c('basic solution', 'solution with corridors'))
# 
## ---------------------------------------------