Boundary matrix

Generate a boundary matrix describing the shared and exposed edges of planning units.

boundary_matrix(x) # Raster

boundary_matrix(x) # SpatialPolygons

boundary_matrix(x) # SpatialLines

boundary_matrix(x) # SpatialPoints

Arguments

x

Raster-class, SpatialLines-class, or SpatialPolygons-class object. If x is a Raster-class object then it must have only one layer.

...

not used.

Value

Matrix{dsCMatrix-class} object.

Details

This function returns a dsCMatrix-class symmetric sparse matrix. Cells on the off-diagonal indicate the length of the shared boundary between two different planning units. Cells on the diagonal indicate length of a given planning unit's edges that have no neighbors (eg. for edges of planning units found along the coastline). This function assumes the data are in a coordinate system where Euclidean distances accurately describe the proximity between two points on the earth. Thus spatial data in a longitude/latitude coordinate system (aka http://spatialreference.org/ref/epsg/wgs-84/) should be reprojected to another coordinate system before using this function.