Run calculations

Execute preliminary calculations in a conservation problem and store the results for later reuse. This function is useful when creating slightly different versions of the same conservation planning problem that involve the same constraints, because means that the same calculations will not be run multiple times.

run_calculations(x)

Arguments

x

ConservationProblem object

Value

None.

Details

This function is used for the effect of modifying the input ConservationProblem object. As such, it does not return anything. To use this function with pipe operators, use the %T>% operator and not the %>% operator.

See also

tee.

Examples

# Let us imagine a scenario where we wanted to understand the effect of 
# setting different targets on our solution.

# create a conservation problem with no targets
p <- problem(sim_pu_raster, sim_features) %>%
  add_minimum_set_objective() %>%
  add_boundary_constraint(10, 0.5)

# create a copies of p and add targets
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# now solve each of the different problems and record the time spent 
# solving them
s1 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [1e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [2e+00, 7e+00]
#> Found heuristic solution: objective 3371.3
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 1.808671e+03, 211 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1808.67141    0   15 3371.29701 1808.67141  46.4%     -    0s
#> H    0     0                    2010.0617996 1808.67141  10.0%     -    0s
#> H    0     0                    1836.5292052 1808.67141  1.52%     -    0s
#> 
#> Explored 0 nodes (211 simplex iterations) in 0.04 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 3: 1836.53 2010.06 3371.3 
#> Pool objective bound 1808.67
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.836529205214e+03, best bound 1.808671409851e+03, gap 1.5169%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [1e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [5e+00, 1e+01]
#> Found heuristic solution: objective 6216.37
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 3.625936e+03, 210 iterations, 0.01 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3625.93604    0   16 6216.37471 3625.93604  41.7%     -    0s
#> H    0     0                    3687.7793461 3625.93604  1.68%     -    0s
#> 
#> Explored 0 nodes (210 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 3687.78 6216.37 
#> Pool objective bound 3625.94
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 3.687779346069e+03, best bound 3.625936037324e+03, gap 1.6770%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [1e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [7e+00, 2e+01]
#> Found heuristic solution: objective 8468.96
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 5.452273e+03, 191 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5452.27334    0   17 8468.96213 5452.27334  35.6%     -    0s
#> H    0     0                    5894.0843471 5452.27334  7.50%     -    0s
#> 
#> Explored 0 nodes (191 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 5894.08 8468.96 
#> Pool objective bound 5452.27
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 5.894084347075e+03, best bound 5.452273338614e+03, gap 7.4958%

# This approach is inefficient. Since these problems all share the same 
# planning units it is actually performing the same calculations three times.
# To avoid this, we can use the "run_calculations" function before creating 
# the copies. Normally, R runs the calculations just before solving the 
# problem

# recreate a conservation problem with no targets and tell R run the
# preliminary calculations. Note how we use the \%T>\% operator here.
p <- problem(sim_pu_raster, sim_features) %>%
  add_minimum_set_objective() %>%
  add_boundary_constraint(10, 0.5) %T>%
  run_calculations()

# create a copies of p and add targets just like before
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# solve each of the different problems and record the time spent 
# solving them
s2 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [1e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [2e+00, 7e+00]
#> Found heuristic solution: objective 3371.3
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 1.808671e+03, 211 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1808.67141    0   15 3371.29701 1808.67141  46.4%     -    0s
#> H    0     0                    2010.0617996 1808.67141  10.0%     -    0s
#> H    0     0                    1836.5292052 1808.67141  1.52%     -    0s
#> 
#> Explored 0 nodes (211 simplex iterations) in 0.03 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 3: 1836.53 2010.06 3371.3 
#> Pool objective bound 1808.67
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.836529205214e+03, best bound 1.808671409851e+03, gap 1.5169%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [1e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [5e+00, 1e+01]
#> Found heuristic solution: objective 6216.37
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 3.625936e+03, 210 iterations, 0.01 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3625.93604    0   16 6216.37471 3625.93604  41.7%     -    0s
#> H    0     0                    3687.7793461 3625.93604  1.68%     -    0s
#> 
#> Explored 0 nodes (210 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 3687.78 6216.37 
#> Pool objective bound 3625.94
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 3.687779346069e+03, best bound 3.625936037324e+03, gap 1.6770%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [1e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [7e+00, 2e+01]
#> Found heuristic solution: objective 8468.96
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 5.452273e+03, 191 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5452.27334    0   17 8468.96213 5452.27334  35.6%     -    0s
#> H    0     0                    5894.0843471 5452.27334  7.50%     -    0s
#> 
#> Explored 0 nodes (191 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 5894.08 8468.96 
#> Pool objective bound 5452.27
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 5.894084347075e+03, best bound 5.452273338614e+03, gap 7.4958%

# now lets compare the times
print(s1) # time spent without running preliminary calculations

#>    user  system elapsed 
#>   1.254   0.197   1.466 

print(s2) # time spent after running preliminary calculations

#>    user  system elapsed 
#>   1.067   0.185   1.265 

# As we can see, we can save a lot of time by running the preliminary 
# calculations before making copies of the problem with slightly
# different constraints.