Problem solvers

Specify the software and configurations used to solve a conservation planning problem. Below is a list of different solvers that can be added to a ConservationProblem-class object.

add_gurobi_solver(x, gap = 0.1, time_limit = .Machine$integer.max,
  presolve = 2, threads = 1, first_feasible = 0)

add_lpsymphony_solver(x, gap = 0.1, time_limit = -1, verbosity = 1,
  first_feasible = 0)

add_rsymphony_solver(x, gap = 0.1, time_limit = -1, first_feasible = 0,
  verbosity = 1)

add_default_solver(x)

Arguments

x

ConservationProblem-class object.

gap

numeric relative gap to optimality. The optimizer will terminate when the difference between the upper and lower objective function bounds is less than the gap times the upper bound. For example, a value of 0.01 will result in the optimizer stopping when the difference between the bounds is 1 percent of the upper bound.

time_limit

numeric time limit in seconds to run the optimizer. The solver will return the current best solution when this time limit is exceeded.

presolve

integer number indicating how intensively the solver should try to simplify the problem before solving it. The default value of 2 indicates to that the solver should be very aggressive in trying to simplify the problem.

threads

integer number of threads to use for the optimization algorithm. The default value of 1 will result in only one thread being used.

first_feasible

logical should the first feasible solution be be returned? If first_feasible is set to TRUE, the solver will return the first solution it encounters that meets all the constraints, regardless of solution quality. Note that the first feasible solution is not an arbitrary solution, rather it is derived from the relaxed solution, and is therefore often reasonably close to optimality.

verbosity

integer how verbose should the solver be when reporting progress on solving the problem?

Details

add_gurobi_solver

http://gurobi.com is a state-of-the-art commercial optimization software with an R package interface. It is by far the fastest of the solvers available in this package, however, it is also the only one that isn't free. That said, free academic licenses are available. The gurobi package is distributed with the Gurobi software suite. This solver uses the gurobi package to solve problems.

add_rsymphony_solver

https://projects.coin-or.org/SYMPHONY is an open-source integer programming solver that is part of the Computational Infrastructure for Operations Research (COIN-OR) project, an initiative to promote development of open-source tools for operations research (a field that includes linear programming). The Rsymphony package provides an interface to COIN-OR and is available on CRAN. This solver uses the Rsymphony package to solve problems.

add_lpsymphony_solver

The lpsymphony package provides a different interface to the COIN-OR software suite. Unlike the Rsymhpony package, the lpsymphony package is distributed through http://bioconducto/packages/release/bioc/html/lpsymphony.html. On Windows and Mac, lpsymphony may be easier to may be easier to install. This solver uses the lpsymphony package to solve.

default_solver

This solver uses the best software installed currently installed on the system.

Examples

# create basic problem with defaults
p <- problem(sim_pu_raster, sim_features) %>%
  add_minimum_set_objective() %>%
  add_relative_targets(0.1)

# add rsymphony solver with default settings
p %>% add_rsymphony_solver()

#> Conservation Problem
#>   planning units: RasterLayer (90 units)
#>   cost:           min: 190.13276, max: 215.86384
#>   features:       layer.1, layer.2, layer.3, layer.4, layer.5
#>   objective:      Minimum set objective 
#>   targets:        Relative targets [targets (min: 0.1, max: 0.1)]
#>   decision:       default
#>   constraints:    <none>
#>   solver:         Rsymphony [verbosity (1), gap (0.1), time_limit (-1), first_feasible (0)]

# add rsymphony solver with custom settings
p %>% add_rsymphony_solver(gap=0.05, time_limit=100)

#> Conservation Problem
#>   planning units: RasterLayer (90 units)
#>   cost:           min: 190.13276, max: 215.86384
#>   features:       layer.1, layer.2, layer.3, layer.4, layer.5
#>   objective:      Minimum set objective 
#>   targets:        Relative targets [targets (min: 0.1, max: 0.1)]
#>   decision:       default
#>   constraints:    <none>
#>   solver:         Rsymphony [verbosity (1), gap (0.05), time_limit (100), first_feasible (0)]

## Not run: ------------------------------------
# # add gurobi solver
# p %>% add_gurobi_solver(gap=0.1, presolve=2, time_limit=100)
# 
# # add lpsolver solver
# p %>% add_lpsymphony_solver(gap=0.1, time_limit=100)
# 
## ---------------------------------------------