Run calculations

Execute preliminary calculations in a conservation problem and store the results for later reuse. This function is useful when creating slightly different versions of the same conservation planning problem that involve the same constraints, because means that the same calculations will not be run multiple times.

run_calculations(x)

Arguments

x

ConservationProblem-class object

Value

None.

Details

This function is used for the effect of modifying the input ConservationProblem-class object. As such, it does not return anything. To use this function with pipe operators, use the %T>% operator and not the %>% operator.

See also

tee.

Examples

# Let us imagine a scenario where we wanted to understand the effect of 
# setting different targets on our solution.

# create a conservation problem with no targets
p <- problem(sim_pu_raster, sim_features) %>%
  add_minimum_set_objective() %>%
  add_boundary_constraint(10, 0.5)

# create a copies of p and add targets
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# now solve each of the different problems and record the time spent 
# solving them
s1 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [3e+00, 7e+00]
#> Found heuristic solution: objective 2166.76
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 1.744084e+03, 188 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1744.08391    0    9 2166.75691 1744.08391  19.5%     -    0s
#> H    0     0                    1811.1737516 1744.08391  3.70%     -    0s
#> 
#> Explored 0 nodes (188 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 1811.17 2166.76 
#> Pool objective bound 1744.08
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.811173751556e+03, best bound 1.744083913573e+03, gap 3.7042%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [6e+00, 1e+01]
#> Found heuristic solution: objective 4176.88
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 3.523702e+03, 191 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3523.70224    0   23 4176.87913 3523.70224  15.6%     -    0s
#> H    0     0                    3616.3443058 3523.70224  2.56%     -    0s
#> 
#> Explored 0 nodes (191 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 3616.34 4176.88 
#> Pool objective bound 3523.7
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 3.616344305786e+03, best bound 3.523702241788e+03, gap 2.5618%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [8e+00, 2e+01]
#> Found heuristic solution: objective 6216.37
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 5.320557e+03, 186 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5320.55718    0    8 6216.37471 5320.55718  14.4%     -    0s
#> H    0     0                    5451.4066953 5320.55718  2.40%     -    0s
#> 
#> Explored 0 nodes (186 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 5451.41 6216.37 
#> Pool objective bound 5320.56
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 5.451406695283e+03, best bound 5.320557179926e+03, gap 2.4003%

# This approach is inefficient. Since these problems all share the same 
# planning units it is actually performing the same calculations three times.
# To avoid this, we can use the "run_calculations" function before creating 
# the copies. Normally, R runs the calculations just before solving the 
# problem

# recreate a conservation problem with no targets and tell R run the
# preliminary calculations. Note how we use the \%T>\% operator here.
p <- problem(sim_pu_raster, sim_features) %>%
  add_minimum_set_objective() %>%
  add_boundary_constraint(10, 0.5) %T>%
  run_calculations()

# create a copies of p and add targets just like before
p1 <- p %>% add_relative_targets(0.1)
p2 <- p %>% add_relative_targets(0.2)
p3 <- p %>% add_relative_targets(0.3)

# solve each of the different problems and record the time spent 
# solving them
s2 <- system.time({solve(p1); solve(p2); solve(p3)})

#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [3e+00, 7e+00]
#> Found heuristic solution: objective 2166.76
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 1.744084e+03, 188 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 1744.08391    0    9 2166.75691 1744.08391  19.5%     -    0s
#> H    0     0                    1811.1737516 1744.08391  3.70%     -    0s
#> 
#> Explored 0 nodes (188 simplex iterations) in 0.01 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 1811.17 2166.76 
#> Pool objective bound 1744.08
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 1.811173751556e+03, best bound 1.744083913573e+03, gap 3.7042%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [6e+00, 1e+01]
#> Found heuristic solution: objective 4176.88
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 3.523702e+03, 191 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 3523.70224    0   23 4176.87913 3523.70224  15.6%     -    0s
#> H    0     0                    3616.3443058 3523.70224  2.56%     -    0s
#> 
#> Explored 0 nodes (191 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 3616.34 4176.88 
#> Pool objective bound 3523.7
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 3.616344305786e+03, best bound 3.523702241788e+03, gap 2.5618%
#> Optimize a model with 295 rows, 235 columns and 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Coefficient statistics:
#>   Matrix range     [2e-01, 1e+00]
#>   Objective range  [1e+00, 2e+02]
#>   Bounds range     [1e+00, 1e+00]
#>   RHS range        [8e+00, 2e+01]
#> Found heuristic solution: objective 6216.37
#> Presolve time: 0.01s
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> Variable types: 0 continuous, 235 integer (235 binary)
#> Presolved: 295 rows, 235 columns, 1030 nonzeros
#> 
#> 
#> Root relaxation: objective 5.320557e+03, 186 iterations, 0.00 seconds
#> 
#>     Nodes    |    Current Node    |     Objective Bounds      |     Work
#>  Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
#> 
#>      0     0 5320.55718    0    8 6216.37471 5320.55718  14.4%     -    0s
#> H    0     0                    5451.4066953 5320.55718  2.40%     -    0s
#> 
#> Explored 0 nodes (186 simplex iterations) in 0.02 seconds
#> Thread count was 1 (of 2 available processors)
#> 
#> Solution count 2: 5451.41 6216.37 
#> Pool objective bound 5320.56
#> 
#> Optimal solution found (tolerance 1.00e-01)
#> Best objective 5.451406695283e+03, best bound 5.320557179926e+03, gap 2.4003%

# now lets compare the times
print(s1) # time spent without running preliminary calculations

#>    user  system elapsed 
#>   1.229   0.174   1.419 

print(s2) # time spent after running preliminary calculations

#>    user  system elapsed 
#>   1.025   0.174   1.207 

# As we can see, we can save a lot of time by running the preliminary 
# calculations before making copies of the problem with slightly
# different constraints.